Linkify reqs table; remove duplicate descriptions.
In the "Level requirements" section, add links in the table from each
requirement the corresponding entry in the detailed source/build/common
doc, and remove the duplicate "summary" descriptions below it. This
avoids confusion and the need to keep the two in sync, at the cost of
making it slightly harder to skim. This seems like a worthwhile
tradeoff.

Also convert the table to pure Markdown, for two reasons:
- To make the table easier maintain.
- To allow us links to work in both GitHub and GitHub Pages. This is
  achieved by the jekyll-relative-links plugin, which only works with
  markdown links, not HTML links.

The downside to converting to Markdown is that we can't use rowspan and
colspan. This isn't too bad of a hit to readability.