Properly indent HTML requirements table.
Remove all the indentation from the requirements table to avoid
confusion and mistakes. Previously we indented each level with one
space, which happened to work because we never used more than three
spaces. But this is misleading because four or more spaces are
interpreted as a code block. Instead, do not indent the HTML table at
all, which is what the CommonMark specification recommends.

Also remove the </tr> tags because they are not needed.