Add slsa v0.2 provenance spec cue validator
This file can be used with cue cli tools to validate attestations
against provenance spec like:

`cue vet provenance-2.0.cue attestation.json`

It will return an error if it doesn't match the spec. This also
opens up the ability to use CUE for codegen for various tools down
the line allowing us to create a canonical validator and tools
written in other languages can just use the CUE spec for both
validation as well as generation of attestations that meet the spec.