README: Add steering committee list, fix GH alias.
Add the newly creating SLSA steering committee mailing list to the
README file to allow people to easily contact the steering committee.
This is not on the slsa.dev site since we'd prefer users to go through
public channels.

Also fix the GitHub team name (no "teams/") in the README.

Signed-off-by: Mark Lodato <lodato@google.com>