Switch from Google Analytics to GoatCounter.
It ends up that the "denied consent" mode doesn't work at all. It sends
the data to Google but you can't actually view it on Google Analytics.
Right now it's showing our site as having no traffic.

The only free, cloud-hosted, GDPR-compliant substitute I've found is
GoatCounter. The site is super basic (see https://slsa.goatcounter.com)
but it's good enough. If we want something better, we'll need to either
pay or self-host.

Signed-off-by: Mark Lodato <lodato@google.com>