Simplify the specifications list template.
Similar to e56c388 (#270), but for the list of subpages in
/spec/v0.1/index. This change replace the "automatic" code with a manual
list, which is much easier to understand and maintain.

No content change. The only HTML change is to switch from absolute to
relative links and to eliminate empty `<li>` elements.

NOTE: Neither the original nor new code correctly handled displaying
more than one specification version. This commit adds a TODO making it
clear that it doesn't work.

Signed-off-by: Mark Lodato <lodato@google.com>