Document a SLSA verification summary attestation.
Fixes #107 by creating an attestation that indicates a `verifier`
has determined that the specified `subject` artifact(s) meets the
indicated SLSA level.

In addition this attestation also indicates the minimum aggregate
SLSA level met by the dependencies used to build the artifact
which can help to address #61.

This leaves a number of things up to the user:
* What it means to evaluate an artifact against policy. (See #46)
* How to communicate the attestations required to create this
    attestation.
* How to debug failures

I borrowed heavily from slsa.dev/provenance in terms of
formatting/documentation.

This attestation is based off or work done by @AdamZWu.

Signed-off-by: Tom Hennen <tomhennen@google.com>