Move "Provenance" under "specifications" menu and add "VSA"
It's a bit confusing why "Provenance" gets its own top-level menu,
especially since it too is a specification, so now all specs are under
the one menu.

Insert a horizontal rule to differentiate between the SLSA ladder and
the attestations formats. Without that, the organization of the menu is
a bit unclear.

Signed-off-by: Mark Lodato <lodato@google.com>