Rename _spec to spec (internal cleanup)
Previously the source files for `/spec/...` were found in the
`docs/_spec/...` directory because it was a Jekyll collection. This was
confusing and made it a little harder to edit. The reason it was this
way was because it was a Jekyll collection, but we never took advantage
of it being a collection.

Now it is just a normal directory, `docs/spec/...`. This makes it easier
to edit.

Signed-off-by: Mark Lodato <lodato@google.com>
Signed-off-by: John Osborne <johnfosborneiii@gmail.com>