Allow hidden versions of spec and fix page link
Allow specifications to be "hidden" so that they don't show up in the
version picker by default.

Also fix a bug where the link to the page did not actually work if the
slugified title did not match the filename.

This moves the list of versions out of _config.yml into a data file,
which is easier to edit and more flexible. A future commit will lean up
old uses of the _config.yml variables.

Signed-off-by: Mark Lodato <lodato@google.com>
Signed-off-by: John Osborne <johnfosborneiii@gmail.com>