Update Slack link to the OpenSSF invite page
The existing links to the Slack workspace were asking new users to login
without an option to create an account. The openssf.org site links to
slack.openssf.org as a landing page for existing users to login and new
users to create accounts -- this patch changes to use the same URL for
Slack.

Signed-off-by: Joshua Lock <jlock@vmware.com>