SLSA v1.0: delete most of index.md
The previous content on spec/*/index.md was redundant with other pages
and fell into an awkward middle ground: it was slightly more detailed
than the top-level splash page but much less detailed than the
individual specification pages. Furthermore, the main thing that people
wanted on this page was the table of contents, but that was all the way
at the bottom.

Instead of keeping it up-to-date for v1.0, let's drop it entirely. Now
the page contains only a table of contents plus a short blurb informing
readers about versioning of the spec and the presence of other
specifications, which is a common point of confusion.

Note: The `<span>` tags are to used to direct the reader to the version
chooser: bottom of page on mobile and right on desktop. You can test
this by resizing your window or using the Chrome developer tools.

Signed-off-by: Mark Lodato <lodato@google.com>