Layer "draft" banner under the nav menu
Previously the "draft" banner was on top of the nav menu, which made it
impossible to navigate unless you dismissed the banner. We don't
actually need a z-index on the banner since it's defined earlier in the
document than the rest of the page, but the nav menu needs a z-index to
make it appear above the banner.

Signed-off-by: Mark Lodato <lodato@google.com>