Move faq to within the versioned directory
Previously the FAQ was unversioned, for reasons I cannot remember. This
is a problem for v1.0 because the naming of the levels changed, making
the FAQ version-specific.

To fix this, this commit copies faq.md to v0.1 and v1.0 (without
modification) and then updates the original location to redirect to v0.1
(since that is the latest published version).

Signed-off-by: Mark Lodato <lodato@google.com>