SLSA v1.0: re-add build reqs from old SLSA 4
The previous commit deleted all of the SLSA 4 Build requirements because
they are undefined in v1.0. However, we may want to tweak the wording
over time, so it would be valuable to have a copy in v1.0 that we can
evolve, even if it is not yet a formal requirement. For example, we
likely need to better explain what the purpose of reproducibility is.

This commit adds back the text verbatim, but hidden under a collapsed
`<details>` element and annotated as "(draft)". This way we can convey
future direction while making it clear that the wording is not final.

The only change was to move the definition of "immutable reference" down
to the one requirement that used the term (Hermetic).

Signed-off-by: Mark Lodato <lodato@google.com>