Provenance v1.0: initial draft (#525)
This is an initial draft v1.0 for the provenance format. It attempts to address
the major sources of confusion with v0.2, particularly the separation of
concerns between top-level external inputs, resolved dependencies, and per-run
details.

There are still many outstanding TODOs and unresolved design decisions. The idea
is to submit this now so that we can more easily get feedback on the schema and
substance, then send further PRs to address the remaining TODOs and RFCs.

Signed-off-by: Mark Lodato <lodato@google.com>