Provenance v1: allow arbitrary JSON for params
Based on feedback, generalize `externalParameters` and
`systemParameters` to allow arbitrary JSON objects rather than only
specific types. This makes it easier to represent real-world systems. It
also avoids the needs for the extra layer of `artifactRef`,
`scalarValue`, etc., which was awkward.

This necessitated the removal of the "artifact reference" type of
parameter, which combined the actual parameter (usually represented as a
URI) with the resolved digest of the artifact that the parameter pointed
to. Now the resolved digest is moved to `resolvedDependencies` (with the
URI repeated there to allow joining with the parameters.) The impetus
for the removal was because there was no good way to represent the
concept of an artifact reference parameter, but upon further reflection,
the removal actually seems to simplify things.

Also expand the guidance for BuildDefinition.

Signed-off-by: Mark Lodato <lodato@google.com>