Fix switching between provenance spec versions
The v1 spec is in a child directory, whereas earlier (0.1 and 0.2)
versions are direct children of the provenance folder. This means that
when the versions-dropdown replaces the v1 version part with 0.1 or 0.2 in
the replacement url includes a trailing '/'. Stripping the trailing url
for all items in the version selector fixes navigating between provenance
specification versions.

Signed-off-by: Joshua Lock <jlock@vmware.com>