Provenance v1: add verification, clarify model
This explains how to verify provenance without expanding the
requirements of the SLSA Build levels. As discussed at the SLSA
Specification SIG meeting on 2023-02-13, adding verification to the
existing levels seems like it adds too much complexity. This commit
tries to strike the balance of still describing the concept without
making it a SLSA requirement.

Also clarify the model:

- Diagram:
  - Simplify the diagram's colors (orange = "to be verified", green =
    "trusted", remove red).
  - Mark the build cache as green since it's covered by the SLSA Build
    requirement, thus trusted.
  - Add a legend (#593.)
  - Use the same text color as the website.
- Better explain what `buildType` means.
- Clarify communication with caches and the platform.

Signed-off-by: Mark Lodato <lodato@google.com>