Fix version selector
The selector currently offers to switch between versions even when the
page does not exist in the other versions, which leads to a 404. This
change fixes that by extending the versions.yaml file to list the
pages that exist in each version and having the selector only offer to
switch between versions in which the page exists.

To replicate the problem simply go to: https://slsa.dev/spec/v1.0/principles
then select "Version 0.1" -> 404

With this fix, only version 1.0 is offered in the selector.

Signed-off-by: Arnaud J Le Hors <lehors@us.ibm.com>