Move draft requirements to future directions page
Draft requirements which were removed from SLSA during the 1.0 cycle by
un-defining Build L4 and removing the source track were left in the
requirements and levels pages as an indicator of future direction.

However, their presence in the requirements page has been noted as a
source of confusion, particularly for new readers.

In order to improve clarity of the requirements while maintaining open
communication of potential future changes, move the draft requirements to
a new "future directions" page.

Fixes: #605

Signed-off-by: Joshua Lock <jlock@vmware.com>