v1-rc1 [1/2]: create copies of files, unmodified
Create a v1-rc1 snapshot of the spec, provenance, and
github-actions-workflow. This is the first commit of two. It just
creates pure file copies without any file modifications in order to make
tracking history easier. It is not a coherent state since the -rc1
versions still link to the non-rc1 files. The next commit will fix those
links.

Signed-off-by: Mark Lodato <lodato@google.com>