Fix pagination
With this change the titles of the pages used in the prev/next
pagination links are retrieved directly from the pages themselves so
that they don't need to be repeated in the front matter, eliminating
a source of error and simplifying maintenance.

It also fixes a broken pagination link and onepage v1.0 (although not
completely).

Signed-off-by: Arnaud J Le Hors <lehors@us.ibm.com>