Make the draft banner more useful.
Most readers probably don't want to be taken to the latest approved
version, but they might want to select a release candidate (or even know
that a release candidate exists). So refer them to the version selector
rather than adding a link.

Also display what version it is a draft of, to make it more clear to
readers.

Signed-off-by: Mark Lodato <lodato@google.com>