Color nav bar section titles differently from links
Previously all nav bar text entries were rendered identically, whether
or not it was clickable. (Section titles are not clickable.) This made
it a little hard to read.

Now (non-clickable) section titles are rendered differently from
(clickable) links. The new color is light green, the same color as the
hover text. This seems like a decent compromise to me. I experimented
with bold, small caps, and others, but a slight change in color seemed
the most subtle and useful.

Signed-off-by: Mark Lodato <lodato@google.com>