Remove references to expectations from levels.md
de5117ad added expectations and verification to the build track in
levels.md, but since then we have decided to split that into a separate
concept ("Verifying Artifacts"). However, the old language remained in
levels.md.

This change addresses those remnants by explaining that the build level
is purely about the provenance, but mentioning that the primary purpose
is to enable verification and matching against expectations. Hopefully
this is the right balance to strike.

Also removes a now-irrelevant TODO.

Fixes #729.
Fixes #735.

Signed-off-by: Mark Lodato <lodato@google.com>