Nav bar: use underline instead of color for hover
The last commit changed the nav bar to use the same color for
hover/focus as non-clickable titles.

This commit changes hover/focus to use underline instead of color
change. This bring makes it visually different from section titles and
brings it in consistency with the body (which also use hover/focus
underlines).

Signed-off-by: Mark Lodato <lodato@google.com>