Provenance v1: re-add `completeness` and `reproducibility`
These were removed from an earlier draft of v1 when we made it implicit
from `builder.id`, but there has been feedback that it is valuable even
if implicit. Now the values are enums (for extensibility) and are
restricting to being strictly stronger than the implicit value from
`builder.id`.

Signed-off-by: Mark Lodato <lodato@google.com>