Remove final "RFC" comment in provenance.md
The last TODO or RFC comment was asking about adding a "tenant" field. I
had anticipated sending out a PR to add such a field before RC2 but I
didn't get it out in time. Let's remove the RFC since we're no longer
asking for comments on this. Users who want this can file an issue.

Signed-off-by: Mark Lodato <lodato@google.com>