css: prevent "draft" banner from occluding :target
This is a follow-on to 9492b35 (#811).

When loading a link to a section within the page, add enough scroll
padding to prevent the header from overlapping with the target.

While 9492b35 set the `scroll-padding-top` of `<html>` statically, this
change sets the `scroll-margin-top` of all clickable elements
dynamically based on whether or not the banner is visible. It's not
possible to dynamically set the style of `<html>` based on the state of
a checkbox (inside the `<html>`) since there is no CSS parent selector.

Signed-off-by: Mark Lodato <lodato@google.com>