impl: fix "view source" link
The "View source on GitHub" link in the footer broke in the switch to
Netlify due to the lack of `site.github.source`. (These links have
always been broken on Netlify deploy previews.)

To fix:

- Hard-code the source path (`/docs`) since this is static.
- Link to the specific commit rather than the branch. This is actually
  better for the user because it is the precise source code that was
  used to build, rather than whatever is on the latest branch.

Now it works in both production and in deploy previews. Deploy previews
still link to the main slsa-framework/slsa repository, but the commit
URL still works even if it's from a fork. (I'd have preferred to link to
the fork, but I couldn't figure out how to do that.)

Other changes:

- Append `?plain=1` to the URL so that we link to the actual Markdown
  source rather than a rendered preview.
- Hard-code the github repo link in the header, since that is also
  static. This way the site fails gracefully in the event that the git
  metadata breaks in the future.

Fixes #922.

Signed-off-by: Mark Lodato <lodato@google.com>