impl: relax review requirements during Draft stage
As discussed at the 2023-08-21 specification meeting, we do not want to
require a 72h waiting period + 3 reviewers for changes to Draft
specifications, since that will make the iteration speed painfully slow.
Draft specifications are clearly marked as unapproved, and there is
still a Candidate approval process to review the whole spec. Therefore,
change CONTRIBUTING to note that Draft versions are manager discretion
regarding waiting period and number of reviewers.

Fixes #953.

Signed-off-by: Mark Lodato <lodato@google.com>