impl: Create new project lifecycle process
As discussed in the 2023-10-09 specification meeting, create a new
"Project lifecycle" for major projects (new version, new track, etc).
The main idea is to assign a "project shepherd" to oversee the project.
This person is responsible for moving the project along. The reason for
creating this is that many projects seem to stagnate without a single
responsible person.

This commit populates the initial list of projects as per our discussion
at the meeting.

Signed-off-by: Mark Lodato <lodato@google.com>