editorial: Clarify the requirements for self-hosted runners on provenance
Resolves: #966

Some CI systems allow for users to configure self-hosted runner
environments for perform builds and CI analysis. While both the build
platform and the self-hosted runners have the ability to affect the
build for the resulting artifact, the SLSA Build requirements do not
need to be imposed on both systems.

This addition to the FAQ is a clarification of the requirements as they
relate to the generation of the provenance.

Signed-off-by: arewm <arewm@users.noreply.github.com>