impl: use consistent Markdown sytnax for tables (#1024)
Address markdownlint warning about inconsistent markdown syntax for
tables. We almost always use a trailing `|`, so add that to the few
places where it's missing. Also disable the lint warning on Liquid
templates that markdownlint fails to parse properly.

This unblocks upgrading to markdownlint-cli v0.39 (#1021).

A future PR might switch to leading-only syntax (#1023), but that is
outside the scope of this PR.

Signed-off-by: Mark Lodato <lodato@google.com>