impl: add instructions on how to diff builds (#1031)
It's not completely straightforward to compare two versions of the built
site (`_site`) since you need to scrub commit IDs and timestamps. This
commit adds a new section to docs/README.md explaining how to do this.

This is useful to do when upgrading to new versions of dependencies
(e.g. #1028) or changing templates, where you want to make sure there
are no unexpected changes.

Ideally this could be done automatically as part of our CI/CD (GitHub
Actions or Netlify), but this is a simple first step.

Signed-off-by: Mark Lodato <lodato@google.com>