rename A and I to be more inclusive
"(A) Untrustworthy producer": expands to not just malicious intent, but
also other things that reduce trust, such as lack of adequate security
controls.

"(I) Use of unintended package": expands to not just accidental use of
the "wrong" package, but also someone intentionally requesting a "wrong"
package, e.g. in a Kubernetes environment.

Signed-off-by: Mark Lodato <lodato@google.com>