content: re-add threats (A) and (B) from v0.1 (#1056)
Re-add the threat descriptions for (A) and (B) from v0.1 to v1.1. They
are copied verbatim, except for removing the `<span>(SLSA 4)</span>`
and `<sup>...</sup>` labels, which no longer apply to the current
version.

Future PRs will tweak the content of these threat descriptions. But for
this PR, I want to just add them as-is so that we can track cleanly in
version control.

Signed-off-by: Mark Lodato <lodato@google.com>