content: Move v1.1 to Candidate Release (CR) (#1117)
This PR proposes to change the status of v1.1 to Candidate Release in
preparation for final publication.

I ought to point out that there is a bunch of VSA related issues that
had been targeted for this release and that have not been addressed. See
Issue #900. However, until someone works on any of these issues there is
no hope of making progress and waiting for these to close will delay
getting 1.1 out indefinitely. Although not ideal I therefore propose to
defer these and publish what we have.

Signed-off-by: Arnaud J Le Hors <lehors@us.ibm.com>