content: draft: source-track: rename SCP to SCS, replace issue links with project board links.  (#1171)
fixes: https://github.com/slsa-framework/slsa/issues/1128

(cleaned up version of https://github.com/slsa-framework/slsa/pull/1166)

This change is in response to the 9.30 slsa specification meeting on
this topic.

A SCS is the full suite of services and ideas relied upon by the
organization to create source revisions.

VCS stuff should mostly fall out of the discussion

Repositories can be used as the concept used when we need to talk about
authN and authZ w.r.t. authentic contributions.