content: draft: Clarify branch protection in source track (#1175)
Re: https://github.com/slsa-framework/slsa/issues/1136

I'm trying to separate "branch protection" which is typically associated
with SCP features from the property we want for the branches intended
for consumption. I've also applied the SCS terminology we adopted in
another PR.

The framing of how this can be accomplished is still a bit TBD, if we
decide to adopt something like #1142 suggests, we could move the Git +
GitHub/GitLab piece out? That's also where we could discuss server side
pre-receive hooks etc.?