content: draft: Add mitigation for malicious source platform admin. (#1188)
Add mitigation for malicious source platform admin.

We didn't have any guidance for this threat. There are a number of ways
we may be able to address this in the future via the SLSA Source Track
and/or tools like gittuf. However, SLSA doesn't currently address them.
This entire section is already labeled as not being handled by SLSA but
does still include other mitigations.

I'm using the same language we have for "Compromise build platform
admin", which seems like the same sort of threat, and should work 'fine'
until we have something better.

Filed #1187 to track

Fixes #1179.

---------

Signed-off-by: Tom Hennen <tomhennen@google.com>
Signed-off-by: Arnaud J Le Hors <lehors@us.ibm.com>
Co-authored-by: Arnaud J Le Hors <lehors@us.ibm.com>