content: draft: reword "Change management process" requirement
fixes #1139

Per #1139, 'documented process' was somewhat confusing and could
be interpreted as meaning some _prose_ documentation. I think the
real aim is to ensure all the rules for making a change to a
branch were followed.

So I changed the text to talk about those rules, rather than
'documented process'.  I think this is more aligned with what
we're looking for?

Signed-off-by: Tom Hennen <tomhennen@google.com>