Improve strength of Source Level 3
Previously level 3 just required the provenance/attestations to be tamper-resistant
but didn't require those attestations to be created at any particular time.
If an SCS were to create these attestations on-demand it would leave revisions more
susceptible to tampering within the SCS between the time of their production and
the time of the request.

By creating the attestations contemporaneously it reduces the period of time
during which a threat actor would be able to falsify this evidence.

Also changed 'Source Attestations' to 'Source Provenance' to be inline with #1204
where we call it 'Provenance' and not 'Attestations'.

fixes #1216

Signed-off-by: Tom Hennen <tomhennen@google.com>