content: draft: clarify orgs are responsible for what gets expunged (#1252)
The prior language made it sound like the SCS was responsible for only
performing expunging for legal requests. That's not reasonable from a
technical perspective. With this update we clarify that it's the
_organization's_ responsibility to ensure this bar is met.

I had thought about adding a whole section of requirements just for the
organization, but that seems out of scope for this change and it would
be a lot of work. We can still do so if we want, but not here.

fixes #1222

Signed-off-by: Tom Hennen <tomhennen@google.com>