content: draft: Add 'principle' protecting anonymous contribution in SLSA (#1249)
As work on the source track progresses the topic of 'identity' comes up
quite a bit. There has been some confusion about what this means, that
it could be that SLSA intends to require legal identities for all
contributors. That isn't the case.

Many in the open source world prefer to contribute without revealing
their 'real' identities as has been practiced for many years. SLSA does
not intend to change that.

This PR tries to make it clear that SLSA does not require real
identities.

refs #1133

---------

Signed-off-by: Tom Hennen <tomhennen@google.com>
Signed-off-by: Tom Hennen <TomHennen@users.noreply.github.com>
Co-authored-by: Aditya Sirish <8928778+adityasaky@users.noreply.github.com>
Co-authored-by: Andrew McNamara <arewm@users.noreply.github.com>