fix: Correct sidebar links on provenance page
When clicking on Schema.Builder, the anchor would go to the table row
with the builder ID instead of the builder heading. This should
deconflict the anchor points to make the sidebar correctly point to the
right place on the page.

The change also makes the tr ids consistent for the RunDetails section.

Signed-off-by: arewm <arewm@users.noreply.github.com>